refactor(HDOS-499): remove new api folder
