+++ frontend/src/app/enums/base-item-type.ts
@@ -1,3 +1,4 @@
+// TODO rename to TransformationType
 export enum BaseItemType {
   COMPONENT = 'COMPONENT',
   WORKFLOW = 'WORKFLOW'
