+++ frontend/src/app/model/connector.ts
