+++ frontend/src/app/model/link.ts
