+++ frontend/src/app/model/operator.ts
