+++ frontend/src/app/model/vertex.ts
