+++ frontend/src/app/model/workflow-content.ts
