+++ frontend/src/app/store/transformation/transformation.state.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter, EntityState } from '@ngrx/entity';
-import { Transformation } from '../../model/new-api/transformation';
+import { Transformation } from '../../model/transformation';
 
 export const transformationEntityAdapter = createEntityAdapter<Transformation>();
